Docs: Network Device Table and Limitations
This commit adds some documentation to highlight that the network device table
that is printed to the terminal during boot is intended to be debug-level output
rather than user-level information. Additionally, it explains a service ordering
issue that prevents the table from presenting the correct final network state
when using NetworkManager as the renderer on Debian/Ubuntu systems.

This addresses the discussion in GH issue #6161

Fixes: GH-6161